requirements.txt
# Core netunicorn packages
netunicorn-base>=0.3.5
netunicorn-client>=0.3.1
netunicorn-library>=0.3.1

# Jupyter and notebook support
jupyter>=1.1.1
jupyterlab>=4.4.2
notebook>=7.4.2
ipykernel>=6.29.5

# Environment and utility packages
python-dotenv>=1.0.0
returns>=0.20.0
pydantic>=2.0.0
typing-extensions>=4.13.2
cloudpickle>=3.1.1
requests>=2.31.0

# RAG and LLM dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
chromadb>=0.4.22
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Development tools
black>=24.2.0  # code formatting
pytest>=8.0.0  # testing
pytest-cov>=4.1.0  # test coverage
beautifulsoup4>=4.12.2
lxml>=4.9.3 